# server/docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.AIAnalysisResult:
    properties:
      highSeverityPatterns:
        type: integer
      processingTime:
        type: number
      suggestedRules:
        type: integer
      timestamp:
        type: string
      totalPatterns:
        type: integer
    type: object
...
  dto.AttackPatternListResponse:
    properties:
      list:
        description: 直接返回model以避免转换
        items:
          $ref: '#/definitions/model.AttackPattern'
        type: array
      page:
        type: integer
      size:
        type: integer
      total:
        type: integer
    type: object
...
  dto.TriggerAnalysisParams:
    properties:
      anomalyThreshold:
        type: number
      clusteringMethod:
        type: string
      minSamples:
        type: integer
      timeRange:
        type: string
    type: object
...
  model.APIResponse:
    description: API响应的标准格式
    properties:
      code:
        description: HTTP状态码
        example: 200
        type: integer
      data:
        description: 响应数据
      error:
        description: 错误信息(当success=false时)
        example: 参数错误
        type: string
      message:
        description: 响应消息
        example: 操作成功
        type: string
      requestId:
        description: 请求ID，用于跟踪
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      success:
        description: 是否成功
        example: true
        type: boolean
      timestamp:
        description: 响应时间戳
        example: "2023-01-01T12:00:00Z"
        type: string
    type: object
...
host: localhost:2333
info:
  contact:
    email: huahua1319873800@outlook.com
    name: API Support
    url: https://github.com/mingrenya/AI-Waf
  description: MRYa 应用防火墙管理系统 API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://github.com/mingrenya/AI-Waf
  title: MRYa-WAF API
  version: "1.0"
paths:
...
  /api/v1/ai-analyzer/analysis/result:
    get:
      parameters:
      - description: 时间范围
        in: query
        name: timeRange
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AIAnalysisResult'
      summary: 获取AI分析结果
      tags:
      - AI分析器
  /api/v1/ai-analyzer/analyze/patterns:
    post:
      consumes:
      - application/json
      parameters:
      - description: 触发参数
        in: body
        name: request
        schema:
          $ref: '#/definitions/dto.TriggerAnalysisParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.APIResponse'
      summary: 触发AI分析（MCP兼容）
      tags:
      - AI分析器
  /api/v1/ai-analyzer/config:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.AIAnalyzerConfig'
      summary: 获取AI分析器配置
      tags:
      - AI分析器
    put:
      consumes:
      - application/json
      parameters:
      - description: 配置请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AIAnalyzerConfigRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.AIAnalyzerConfig'
      summary: 更新AI分析器配置
      tags:
      - AI分析器
  /api/v1/ai-analyzer/conversations:
    get:
      consumes:
      - application/json
      parameters:
      - description: 关联的攻击模式ID
        in: query
        name: patternId
        type: string
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SuccessResponseNoData'
      summary: 获取MCP对话列表
      tags:
      - AI分析器
  /api/v1/ai-analyzer/conversations/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 对话ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SuccessResponseNoData'
      summary: 删除MCP对话
      tags:
      - AI分析器
    get:
      consumes:
      - application/json
      parameters:
      - description: 对话ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.MCPConversation'
      summary: 获取MCP对话详情
      tags:
      - AI分析器
  /api/v1/ai-analyzer/patterns:
    get:
      consumes:
      - application/json
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: size
        type: integer
      - description: 严重程度
        enum:
        - critical
        - high
        - medium
        - low
        in: query
        name: severity
        type: string
      - description: 攻击类型
        in: query
        name: attackType
        type: string
      - description: 开始时间(RFC3339)
        in: query
        name: startTime
        type: string
      - description: 结束时间(RFC3339)
        in: query
        name: endTime
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AttackPatternListResponse'
      summary: 获取攻击模式列表
      tags:
      - AI分析器
  /api/v1/ai-analyzer/patterns/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 模式ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SuccessResponseNoData'
      summary: 删除攻击模式
      tags:
      - AI分析器
    get:
      consumes:
      - application/json
      parameters:
      - description: 模式ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.AttackPattern'
      summary: 获取攻击模式详情
      tags:
      - AI分析器
  /api/v1/ai-analyzer/rules:
    get:
      consumes:
      - application/json
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: size
        type: integer
      - description: 规则状态
        enum:
        - pending
        - approved
        - rejected
        - deployed
        in: query
        name: status
        type: string
      - description: 规则类型
        enum:
        - modsecurity
        - microrule
        in: query
        name: ruleType
        type: string
      - description: 关联的攻击模式ID
        in: query
        name: patternId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GeneratedRuleListResponse'
      summary: 获取生成规则列表
      tags:
      - AI分析器
  /api/v1/ai-analyzer/rules/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 规则ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SuccessResponseNoData'
      summary: 删除生成规则
      tags:
      - AI分析器
    get:
      consumes:
      - application/json
      parameters:
      - description: 规则ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.GeneratedRule'
      summary: 获取生成规则详情
      tags:
      - AI分析器
  /api/v1/ai-analyzer/rules/{id}/deploy:
    post:
      consumes:
      - application/json
      parameters:
      - description: 规则ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SuccessResponseNoData'
      summary: 部署规则
      tags:
      - AI分析器
  /api/v1/ai-analyzer/rules/pending:
    get:
      consumes:
      - application/json
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SuccessResponseNoData'
      summary: 获取待审核规则列表
      tags:
      - AI分析器
  /api/v1/ai-analyzer/rules/review:
    post:
      consumes:
      - application/json
      parameters:
      - description: 审核请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ReviewRuleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SuccessResponseNoData'
      summary: 审核规则
      tags:
      - AI分析器
  /api/v1/ai-analyzer/stats:
    get:
      consumes:
      - application/json
      parameters:
      - description: 开始时间(RFC3339)
        in: query
        name: startTime
        type: string
      - description: 结束时间(RFC3339)
        in: query
        name: endTime
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AIAnalysisStatsResponse'
      summary: 获取AI分析器统计信息
      tags:
      - AI分析器
  /api/v1/ai-analyzer/suggestions:
    get:
      parameters:
      - description: 状态
        in: query
        name: status
        type: string
      - description: 严重等级
        in: query
        name: severity
        type: string
      - default: 20
        description: 数量
        in: query
        name: limit
        type: integer
      - default: 0
        description: 偏移
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AIRuleSuggestionListResponse'
      summary: 获取AI规则建议列表
      tags:
      - AI分析器
  /api/v1/ai-analyzer/suggestions/{id}/approve:
    post:
      consumes:
      - application/json
      parameters:
      - description: 建议ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SuccessResponseNoData'
      summary: 批准规则建议
      tags:
      - AI分析器
  /api/v1/ai-analyzer/suggestions/{id}/deploy:
    post:
      consumes:
      - application/json
      parameters:
      - description: 建议ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SuccessResponseNoData'
      summary: 部署规则建议
      tags:
      - AI分析器
  /api/v1/ai-analyzer/suggestions/{id}/reject:
    post:
      consumes:
      - application/json
      parameters:
      - description: 建议ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SuccessResponseNoData'
      summary: 拒绝规则建议
      tags:
      - AI分析器
...
securityDefinitions:
  BearerAuth:
    description: 使用 Bearer {token} 格式进行身份验证
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"